designing the cards
