+++ src/components/tasks/createTask/CreateTask.jsx
@@ -125,9 +125,9 @@ const CreateTask = () => {
             required
           >
             <option value=""> </option>
-            <option value="low">Low</option>
-            <option value="medium">Medium</option>
-            <option value="high">High</option>
+            <option value="Low Priority">Low</option>
+            <option value="Medium Priority">Medium</option>
+            <option value="High Priority">High</option>
           </select>
         </div>
